src/backend/CupcakeDias.Shared/Services/Implementations/UserService.cs
using System;
using CupcakeDias.Data;
using CupcakeDias.Data.Entities;
using CupcakeDias.Shared.Services.Interfaces;

namespace CupcakeDias.Shared.Services.Implementations;

public class UserService(CupcakeDiasContext context) : IUserService
{
    private readonly CupcakeDiasContext _context = context;

    public async Task<User> CreateUserAsync(User user)
    {
        _context.Users.Add(user);
        await _context.SaveChangesAsync();
        return user;
    }

    public async Task<User> GetUserByIdAsync(int userId)
    {
        return await _context.Users.FindAsync(userId);
    }

    public async Task<User> GetUserByEmailAsync(string email)
    {
        return await _context.Users.FirstOrDefaultAsync(u => u.Email == email);
    }
}
